added statistics perfomance improvements sound reviewed bugfixes
git-svn-id: https://svn.code.sf.net/p/spacezero/code@41 c47e33d3-9b0f-4996-8ff2-4dadeca765d7

+++ include/functions.h
@@ -61,6 +61,6 @@ void IncTime(void);
 
 void DelCharFromCad(char *cad,char *filter);
 
-int MemUsed(int action,int value);
+int MemUsed(int action,size_t value);
 
 #endif
